+++ src/scrypt_async.h
@@ -13,9 +13,9 @@ public:
     const unsigned int passlen,
     const unsigned char *salt,
     size_t saltlen,
-    uint64_t N,
-    uint64_t r,
-    uint64_t p,
+    unsigned long long N,
+    unsigned long long r,
+    unsigned long long p,
     size_t keylen,
     Nan::Callback *callback
   );
